fix bug: 1. Add "UpdaterSchurVINS" declaration to VioManager.h; 2. Type "MXD" not define in StepHelper.cpp; 3. Add "res, V, V_inv and gv" declaration to UpdaterHelper.h for fix UpdaterSchurVINS.cpp; 4. fix type "Matrix2o3d, Matrix3o6d, Matrix2o6d, Matrix2o3d" not define in UpdaterSchurVINS.cpp; 5. rename "UpdaterSchurVINS.hpp" to "UpdaterSchurVINS.h"
