+++ ov_msckf/src/core/VioManager.h
@@ -48,6 +48,7 @@ namespace ov_msckf {
 class State;
 class StateHelper;
 class UpdaterMSCKF;
+class UpdaterSchurVINS;
 class UpdaterSLAM;
 class UpdaterZeroVelocity;
 class Propagator;
